Refactor `OverridableEnvironment` and `IOEnvironment` into `BuildEnvironment`.
