+++ build_resolvers/pubspec.yaml
@@ -11,7 +11,7 @@ dependencies:
   analyzer: '>=6.9.0 <8.0.0'
   async: ^2.5.0
   build: ^2.4.3-wip
-  build_runner_core: ^8.0.1-wip
+  build_runner_core: ^9.0.0-wip
   collection: ^1.17.0
   convert: ^3.1.1
   crypto: ^3.0.0
