+++ build_runner/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Bump the min SDK to 3.7.0.
 - Use `build_test` 3.0.0.
+- Use `build_runner_core` 9.0.0.
 - Start using `package:build/src/internal.dart'.
 - Refactor `MultiPackageAssetReader` to internal `AssetFinder`.
 - `FinalizedReader` no longer implements `AssetReader`.
