+++ build_runner/pubspec.yaml
@@ -20,7 +20,7 @@ dependencies:
   build_config: ">=1.1.0 <1.2.0"
   build_daemon: ^4.0.0
   build_resolvers: ^2.4.4
-  build_runner_core: ^8.0.1-wip
+  build_runner_core: ^9.0.0-wip
   code_builder: ^4.2.0
   collection: ^1.15.0
   crypto: ^3.0.0
