+++ build_runner_core/lib/build_runner_core.dart
@@ -8,8 +8,6 @@ export 'src/asset/finalized_reader.dart';
 export 'src/asset/reader_writer.dart';
 export 'src/asset/writer.dart';
 export 'src/environment/build_environment.dart';
-export 'src/environment/io_environment.dart';
-export 'src/environment/overridable_environment.dart';
 export 'src/generate/build_directory.dart';
 export 'src/generate/build_result.dart';
 export 'src/generate/build_runner.dart';
