+++ build_test/pubspec.yaml
@@ -12,7 +12,7 @@ dependencies:
   build: ^2.4.3-wip
   build_config: ^1.0.0
   build_resolvers: ^2.4.0
-  build_runner_core: ^8.0.1-wip
+  build_runner_core: ^9.0.0-wip
   convert: ^3.0.0
   crypto: ^3.0.0
   glob: ^2.0.0
